python/requirements.txt
pyobjc>=10.3
ruff>=0.6
